move APIFnDel and ActionDel to action_delete.go
